requirements.txt
torch==2.1.0
torchvision==0.16.0
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
camel-ai[all]
PyMuPDF
groq
python-dotenv
